requirements.txt
distlib ==0.3.6
freezegun
pynsist ==2.8
pyyaml >=6.0.0
wheel
yq >=3.0.0
